main/requirements.txt
asgiref==3.2.7
Django==3.0.4
gunicorn==20.0.4
Pillow==7.1.1
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
